Obligation d'être connecté sur les pages du panneau de gestion

+++ app/Http/Controllers/UtilisateursController.php
@@ -9,6 +9,13 @@ class UtilisateursController extends Controller
 {
     public function liste()
     {
+      if(Auth()->guest())
+      {
+          flash('Vous devez être connecté pour voir cette page')->error();
+
+          return redirect('/connexion');
+      }
+
       $utilisateurs = Utilisateur::all();
 
       return view('utilisateurs', [
@@ -18,6 +25,13 @@ class UtilisateursController extends Controller
     }
     public function voir()
     {
+        if(Auth()->guest())
+        {
+          flash('Vous devez être connecté pour voir cette page')->error();
+
+          return redirect('/connexion');
+        }
+
         $pseudo = request('pseudo');
 
         $utilisateur = Utilisateur::where('pseudo', $pseudo)->first();
@@ -31,8 +45,28 @@ class UtilisateursController extends Controller
 
     public function delete($pseudo)
     {
+        if(Auth()->guest())
+        {
+          flash('Vous devez être connecté pour voir cette page')->error();
+
+          return redirect('/connexion');
+        }
+
         DB::table('utilisateurs')->where('pseudo',$pseudo)->delete();
 
         return redirect('/utilisateurs');
+    }
+    public function update($pseudo)
+    {
+      if(Auth()->guest())
+      {
+          flash('Vous devez être connecté pour voir cette page')->error();
+
+          return redirect('/connexion');
+      }
+
+      DB::table('utilisateurs')->where('pseudo',$pseudo)->update();
+
+      return redirect('/utilisateurs');
     }
 }
